feat: Add mocked post detail apis

+++ src/main/java/com/noair/easip/house/controller/dto/RentDto.java
@@ -5,9 +5,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(description = "집세")
 public record RentDto(
         @Schema(description = "보증금", example = "100000000")
-        Long deposit,
+        Double deposit,
 
         @Schema(description = "월세", example = "500000")
-        Long monthlyRent
+        Double monthlyRent
 ) {
 }
